refactor: packages folder names
