+++ packages/kotti-avatar-group/index.js
@@ -1,4 +1,4 @@
-import KtAvatarGroup from './src/avatarGroup';
+import KtAvatarGroup from './src/avatar-group';
 
 KtAvatarGroup.install = function(Vue) {
 	Vue.component(KtAvatarGroup.name, KtAvatarGroup);
