+++ packages/kotti-avatar-group/package.json
