+++ packages/kotti-avatar-group/src/avatar-group.vue
