+++ packages/kotti-button-group/index.js
@@ -1,4 +1,4 @@
-import KtButtonGroup from './src/buttonGroup';
+import KtButtonGroup from './src/button-group';
 
 KtButtonGroup.install = function(Vue) {
 	Vue.component(KtButtonGroup.name, KtButtonGroup);
